+++ RELEASE_NOTES.rst
@@ -11,9 +11,17 @@ Release Notes
 Upcoming Release
 ================
 
+- Adds option to toggle whether ERA5 downloads are requested in monthly or
+  annual chunks with keyword argument ``cutout.prepare(monthly_requests=True)``.
+  The default is now annual requests. The monthly requests can also be posted
+  concurrently using ``cutout.prepare(monthly_requests=True,
+  concurrent_requests=True)``.
+
 - Improved parallelization of ``atlite.convert.build_line_rating`` by adding
   keyword arguments for ``dask.compute`` (``dask_kwargs={}``) and an option to
-  disable the progressbar (``show_progressbar=True``).
+  disable the progressbar (``show_progress=False``).
+
+- Default to ``show_progress=False`` for performance reasons.
 
 Version 0.2.13
 ==============